split out test helpers
